moved(TextureCache.*): First properties and methods moved to TextureArrayManager
