+++ Assets/UnZENity-Core/Scripts/Caches/TextureCache.cs
@@ -1,16 +1,10 @@
-using System;
 using System.Collections.Generic;
-using System.Diagnostics;
 using System.IO;
-using System.Linq;
-using System.Threading.Tasks;
 using GUZ.Core.Extensions;
-using GUZ.Core.World;
+using GUZ.Core.Manager;
 using JetBrains.Annotations;
 using UnityEngine;
-using UnityEngine.Rendering;
 using ZenKit;
-using Debug = UnityEngine.Debug;
 using Mesh = UnityEngine.Mesh;
 using Object = UnityEngine.Object;
 using Texture = UnityEngine.Texture;
@@ -31,30 +25,18 @@ namespace GUZ.Core.Caches
     /// </summary>
     public static class TextureCache
     {
-        public const int ReferenceTextureSize = 256;
-        public const int MaxTextureSize = 512;
-
-        public static Dictionary<TextureArrayTypes, Texture> TextureArrays { get; } = new();
-        public static List<(MeshRenderer Renderer, WorldData.SubMeshData SubmeshData)> WorldMeshRenderersForTextureArray = new();
+        public static Dictionary<TextureArrayManager.TextureArrayTypes, Texture> TextureArrays { get; } = new();
         public static Dictionary<Mesh, VobMeshData> VobMeshesForTextureArray = new();
 
         private static readonly Dictionary<string, Texture2D> _texture2DCache = new();
-        public static readonly Dictionary<TextureArrayTypes, List<(string PreparedKey, ZkTextureData TextureData)>> TexturesToIncludeInArray = new();
-
-        public enum TextureArrayTypes
-        {
-            Opaque,
-            Transparent,
-            Water
-        }
 
         public class VobMeshData
         {
             public IMultiResolutionMesh Mrm { get; set; }
-            public List<TextureArrayTypes> TextureArrayTypes { get; set; }
+            public List<TextureArrayManager.TextureArrayTypes> TextureArrayTypes { get; set; }
             public List<Renderer> Renderers { get; set; } = new();
 
-            public VobMeshData(IMultiResolutionMesh mrm, List<TextureArrayTypes> textureArrayTypes, Renderer renderer = null)
+            public VobMeshData(IMultiResolutionMesh mrm, List<TextureArrayManager.TextureArrayTypes> textureArrayTypes, Renderer renderer = null)
             {
                 Mrm = mrm;
                 TextureArrayTypes = textureArrayTypes;
@@ -62,25 +44,6 @@ namespace GUZ.Core.Caches
                 {
                     Renderers.Add(renderer);
                 }
-            }
-        }
-
-        [Serializable]
-        public class ZkTextureData
-        {
-            public string Key;
-            public Vector2 Scale;
-            public int Width;
-            public int Height;
-            public int MipmapCount;
-
-            public ZkTextureData(string key, ITexture zkTexture)
-            {
-                Key = key;
-                MipmapCount = zkTexture.MipmapCount;
-                Width = zkTexture.Width;
-                Height = zkTexture.Height;
-                Scale = new Vector2((float)zkTexture.Width / ReferenceTextureSize, (float)zkTexture.Height / ReferenceTextureSize);
             }
         }
 
@@ -154,191 +117,6 @@ namespace GUZ.Core.Caches
             return texture;
         }
 
-        public static void GetTextureArrayIndex(IMaterial materialData, out TextureArrayTypes textureArrayType, out int arrayIndex, out Vector2 textureScale, out int maxMipLevel)
-        {
-            string key = materialData.Texture;
-
-            for (int i = 0; i < Enum.GetValues(typeof(TextureArrayTypes)).Length; i++)
-            {
-                if (!TexturesToIncludeInArray.ContainsKey((TextureArrayTypes)i))
-                {
-                    continue;
-                }
-
-                (string, ZkTextureData) cachedTextureData = TexturesToIncludeInArray[(TextureArrayTypes)i].FirstOrDefault(k => k.PreparedKey == key);
-                if (cachedTextureData != default)
-                {
-                    textureArrayType = (TextureArrayTypes)i;
-
-                    arrayIndex = TexturesToIncludeInArray[textureArrayType].IndexOf(cachedTextureData);
-                    ZkTextureData zkData = TexturesToIncludeInArray[textureArrayType][arrayIndex].TextureData;
-                    maxMipLevel = zkData.MipmapCount - 1;
-                    textureScale = zkData.Scale;
-                    return;
-                }
-            }
-
-            ITexture zenTextureData = ResourceLoader.TryGetTexture(key);
-
-            if (zenTextureData == null)
-            {
-                textureArrayType = default;
-                arrayIndex = -1;
-                textureScale = Vector2.zero;
-                maxMipLevel = 0;
-                return;
-            }
-
-            maxMipLevel = zenTextureData.MipmapCount - 1;
-            UnityEngine.TextureFormat textureFormat = zenTextureData.Format.AsUnityTextureFormat();
-
-            if (materialData.Group == MaterialGroup.Water)
-            {
-                textureArrayType = TextureArrayTypes.Water;
-            }
-            else
-            {
-                textureArrayType = textureFormat == UnityEngine.TextureFormat.DXT1
-                    ? TextureArrayTypes.Opaque
-                    : TextureArrayTypes.Transparent;
-            }
-
-            textureScale = new Vector2((float)zenTextureData.Width / ReferenceTextureSize, (float)zenTextureData.Height / ReferenceTextureSize);
-            if (!TexturesToIncludeInArray.ContainsKey(textureArrayType))
-            {
-                TexturesToIncludeInArray.Add(textureArrayType, new List<(string, ZkTextureData)>());
-            }
-
-            TexturesToIncludeInArray[textureArrayType].Add((key, new ZkTextureData(key, zenTextureData)));
-            arrayIndex = TexturesToIncludeInArray[textureArrayType].Count - 1;
-        }
-
-        public static async Task BuildTextureArrays()
-        {
-            Stopwatch stopwatch = new();
-            stopwatch.Start();
-
-            foreach (TextureArrayTypes textureArrayType in TexturesToIncludeInArray.Keys)
-            {
-                if (Application.isEditor)
-                {
-                    for (int i = 16; i <= 2048; i *= 2)
-                    {
-                        int resCount = TexturesToIncludeInArray[textureArrayType].Where(t => Mathf.Max(t.TextureData.Width, t.TextureData.Height) == i).Count();
-                        Debug.Log($"[{nameof(TextureCache)}] {textureArrayType}: {resCount} textures ({Mathf.RoundToInt(100 * (float)resCount / TexturesToIncludeInArray[textureArrayType].Count)}%) with dimension {i}.");
-                    }
-                }
-
-                // Create the texture array with the max size of the contained textures, limited by the max texture size.
-                int maxSize = Mathf.Min(MaxTextureSize, TexturesToIncludeInArray[textureArrayType].Max(p => p.TextureData.Width));
-                ZkTextureData textureWithMaxAllowedSize = TexturesToIncludeInArray[textureArrayType].Where(t => t.TextureData.Width == maxSize && t.TextureData.Height == maxSize).Select(t => t.TextureData).FirstOrDefault();
-                // Find the max mip depth defined for the max allowed texture size.
-                int maxMipDepth = 0;
-                if (textureWithMaxAllowedSize == null)
-                {
-                    Debug.LogError($"[{nameof(TextureCache)}] No texture with size {maxSize}x{maxSize} was found to determine max allowed mip level. Falling back to texture with highest mip count.");
-                    maxMipDepth = TexturesToIncludeInArray[textureArrayType].Max(t => t.TextureData.MipmapCount);
-                }
-                else
-                {
-                    maxMipDepth = textureWithMaxAllowedSize.MipmapCount;
-                }
-
-                UnityEngine.TextureFormat textureFormat = UnityEngine.TextureFormat.RGBA32;
-                if (textureArrayType == TextureArrayTypes.Opaque)
-                {
-                    textureFormat = UnityEngine.TextureFormat.DXT1;
-                }
-
-                Texture texArray;
-                if (textureArrayType != TextureArrayTypes.Water)
-                {
-                    texArray = new Texture2DArray(maxSize, maxSize, TexturesToIncludeInArray[textureArrayType].Count, textureFormat, maxMipDepth, false, true)
-                    {
-                        filterMode = FilterMode.Trilinear,
-                        wrapMode = TextureWrapMode.Repeat
-                    };
-                }
-                else
-                {
-                    texArray = new RenderTexture(maxSize, maxSize, 0, RenderTextureFormat.ARGB32, maxMipDepth)
-                    {
-                        dimension = TextureDimension.Tex2DArray,
-                        autoGenerateMips = false,
-                        filterMode = FilterMode.Trilinear,
-                        useMipMap = true,
-                        volumeDepth = TexturesToIncludeInArray[textureArrayType].Count,
-                        wrapMode = TextureWrapMode.Repeat
-                    };
-                }
-
-                // Copy all the textures and their mips into the array. Textures which are smaller are tiled so bilinear sampling isn't broken - this is why it's not possible to pack different textures together in the same slice.
-                for (int i = 0; i < TexturesToIncludeInArray[textureArrayType].Count; i++)
-                {
-                    Texture2D sourceTex = TryGetTexture(TexturesToIncludeInArray[textureArrayType][i].PreparedKey, false);
-
-                    int sourceMip = 0;
-                    int sourceMaxDim = Mathf.Max(sourceTex.width, sourceTex.height);
-                    int sourceWidth = sourceTex.width;
-                    int sourceHeight = sourceTex.height;
-                    while (sourceMaxDim > MaxTextureSize)
-                    {
-                        sourceMip++;
-                        sourceMaxDim /= 2;
-                        sourceWidth /= 2;
-                        sourceHeight /= 2;
-                    }
-
-                    for (int mip = sourceMip; mip < sourceTex.mipmapCount; mip++)
-                    {
-                        int targetMip = mip - sourceMip;
-                        for (int x = 0; x < texArray.width / sourceWidth; x++)
-                        {
-                            for (int y = 0; y < texArray.height / sourceHeight; y++)
-                            {
-                                if (texArray is Texture2DArray)
-                                {
-                                    Graphics.CopyTexture(sourceTex, 0, mip, 0, 0, sourceTex.width >> mip,
-                                        sourceTex.height >> mip, texArray, i, targetMip, (sourceTex.width >> mip) * x,
-                                        (sourceTex.height >> mip) * y);
-                                }
-                                else
-                                {
-                                    CommandBuffer cmd = CommandBufferPool.Get();
-                                    RenderTexture rt = (RenderTexture)texArray;
-                                    cmd.SetRenderTarget(new RenderTargetBinding(new RenderTargetSetup(rt.colorBuffer, rt.depthBuffer, targetMip, CubemapFace.Unknown, i)));
-                                    Vector2 scale = new Vector2((float)sourceTex.width / texArray.width, (float)sourceTex.height / texArray.height);
-                                    Blitter.BlitQuad(cmd, sourceTex, new Vector4(1, 1, 0, 0), new Vector4(scale.x, scale.y, scale.x * x, scale.y * y), mip, false);
-                                    Graphics.ExecuteCommandBuffer(cmd);
-                                    cmd.Clear();
-                                    CommandBufferPool.Release(cmd);
-                                }
-                            }
-                        }
-                    }
-
-                    Object.Destroy(sourceTex);
-
-                    if (i % 20 == 0)
-                    {
-                        await Task.Yield();
-                    }
-                }
-
-                TextureArrays.Add(textureArrayType, texArray);
-            }
-
-            // Clear cached texture data to save storage.
-            foreach (List<(string PreparedKey, ZkTextureData Texture)> textureList in TexturesToIncludeInArray.Values)
-            {
-                textureList.Clear();
-                textureList.TrimExcess();
-            }
-
-            stopwatch.Stop();
-            Debug.Log($"Built tex array in {stopwatch.ElapsedMilliseconds / 1000f} s");
-        }
-
         /// <summary>
         /// Unity doesn't want to create mips for DXT1 textures. Recreate them as RGB24.
         /// </summary>
@@ -363,12 +141,6 @@ namespace GUZ.Core.Caches
         {
             TextureArrays.Clear();
             TextureArrays.TrimExcess();
-
-            TexturesToIncludeInArray.Clear();
-            TexturesToIncludeInArray.TrimExcess();
-
-            WorldMeshRenderersForTextureArray.Clear();
-            WorldMeshRenderersForTextureArray.TrimExcess();
         }
 
         private static string GetPreparedKey(string key)
@@ -390,12 +162,6 @@ namespace GUZ.Core.Caches
 
             TextureArrays.Clear();
             TextureArrays.TrimExcess();
-
-            foreach (var textureList in TexturesToIncludeInArray.Values)
-            {
-                textureList.Clear();
-                textureList.TrimExcess();
-            }
         }
     }
 }
